refactor: change post font to PT Serif
I think it reads a little nicer than IBM PS for long-form content and,
with a few letter-spacing / line-spacing tweaks, makes posts look a lot
more readable -- before, they did look somewhat tiring.

This commit also removes the hard-coding of the post's font face.
There are a few more instances of this in the codebase, which will
gradually be removed as I settle on final configurations.